requirements.txt
absl-py==0.10.0
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Flask==1.1.2
folium==0.11.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
